refactor: isolate ui helpers
