+++ src/components/DataUpload.tsx
@@ -2,7 +2,7 @@ import React, { useState, useCallback } from 'react';
 import { useTranslation } from '../hooks/useTranslation';
 import { Button } from '@/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { toast } from 'sonner';
+import { toast } from '@/components/ui/sonner-toast';
 import { 
   MAX_FILE_SIZE, 
   MAX_CSV_ROWS,
